TIP-842: Move group type Controller to Structure
